add TcpModel collector script
This uses tcpdump to capture mysql traffic, which is passed
to pt-tcp-model script from percona toolkit to perform
analysis and print metrics to stdout.

The main advantage is that you can sample query
latency from a running mysql instance with a server
version of 5.5 or below.  5.6 allows for easier collecting
of this kind of data directly from the mysql process.